RAG_inference/requirements.txt

torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
mlflow>=2.9.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
psutil>=5.9.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
